requirements.txt
yt-dlp>=2024.1.0
anthropic>=0.34.0
python-dotenv>=1.0.0
